requirements.txt
scepter>=1.1.0
pycocotools
pyyaml>=5.3.1
scikit-image
torchsde
transformers
scikit-learn
numpy
opencv-python
opencv_transforms>=0.0.6
oss2>=2.15.0
einops
ninja
flash-attn==2.5.8 --use-pep517 --no-build-isolation --verbose
bitsandbytes
diffusers
addict
datasets==3.0.1
# torch==2.4.0+cu121